+++ internal/pkg/upload/upload_test.go
@@ -5,90 +5,6 @@
 package upload
 
 /*
-func TestMaxParallelUploadOpsReached(t *testing.T) {
-	_ = testlog.SetLogger(t)
-	opLimit := 4
-
-	u := New(500, opLimit, 0)
-
-	var err error
-	for i := 0; i < opLimit; i++ {
-		_, err = u.Begin(100, "", "")
-		assert.NoError(t, err)
-	}
-
-	_, err = u.Begin(100, "", "")
-	assert.ErrorIs(t, err, ErrMaxConcurrentUploads)
-}
-
-func TestMaxParallelUploadOpsReleased(t *testing.T) {
-	opLimit := 4
-	u := New(500, opLimit, 0)
-
-	// generate max operations
-	ops := make([]Info, 0, opLimit)
-	for i := 0; i < opLimit; i++ {
-		op, err := u.Begin(100, strconv.Itoa(i), "")
-		require.NoError(t, err)
-		ops = append(ops, op)
-	}
-	// and verify max was reached
-	_, err := u.Begin(100, "X", "")
-	assert.ErrorIs(t, err, ErrMaxConcurrentUploads)
-
-	// finishing an op should release the hold and allow another to begin
-	_, err = u.Complete(ops[0].ID)
-	require.NoError(t, err)
-
-	time.Sleep(5 * time.Millisecond) // occasionally, a little time was required for the change to propagate
-
-	_, err = u.Begin(100, "Y", "")
-	assert.NoError(t, err)
-}
-
-func TestMaxParallelChunks(t *testing.T) {
-	chunkLim := 3
-
-	u := New(104857600, 1, chunkLim)
-
-	// start an operation, that can have more than the test limit chunks
-	op, err := u.Begin(MaxChunkSize*int64(chunkLim+2), "", "")
-	require.NoError(t, err)
-
-	// upload up to the limit chunks, without releasing the request
-	for i := 0; i < chunkLim; i++ {
-		_, err := u.Chunk(op.ID, i)
-		require.NoError(t, err)
-	}
-
-	_, err = u.Chunk(op.ID, chunkLim)
-	assert.ErrorIs(t, err, ErrMaxConcurrentUploads)
-}
-
-func TestMaxParallelChunksReleased(t *testing.T) {
-	chunkLim := 3
-
-	u := New(104857600, 1, chunkLim)
-
-	// start an operation, that can have more than the test limit chunks
-	op, err := u.Begin(MaxChunkSize*int64(chunkLim+2), "", "")
-	require.NoError(t, err)
-
-	// upload up to the limit chunks, without releasing the request
-	chunks := make([]ChunkInfo, 0, chunkLim)
-	for i := 0; i < chunkLim; i++ {
-		info, err := u.Chunk(op.ID, i)
-		require.NoError(t, err)
-		chunks = append(chunks, info)
-	}
-	_, err = u.Chunk(op.ID, chunkLim)
-	assert.ErrorIs(t, err, ErrMaxConcurrentUploads)
-
-	chunks[0].Token.Release()
-
-	_, err = u.Chunk(op.ID, chunkLim)
-	assert.NoError(t, err)
-}
 
 func TestUploadChunkCount(t *testing.T) {
 	tests := []struct {
